HW13/Part2/mysite2/pyproject.toml
[tool.poetry]
name = "my_project"
version = "0.1.0"
description = "MongoDB Redis Quote Caching"
authors = ["Your Name <your-email@example.com>"]

[tool.poetry.dependencies]
psycopg2 = "^2.9"
python = "^3.12"
pymongo = "^3.12"
mongoengine = "^0.20"
redis = "^4.5.1"
redis-lru = "^0.1.2"
dnspython = "^2.7.0"
python-dotenv = "^0.20.0"
pika = "^1.3.2"
django = "^5.1.2"
python-decouple = "^3.8"
dj-database-url = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
